rawjpeg.py does stuff with JPEGs in RAWs. (Just extract for now.)
The idea is to be able to replace the full size JPEG with a higher quality
one without breaking the RAW by moving things. This should be possible as
long as this JPEG is the last thing in the file, as it seems to be in both
PEF and DNG files written by Pentax cameras.